occurrences/requirements.txt
geographiclib>=1.49
google-cloud-firestore>=0.28.0
httplib2>=0.9.2
idigbio>=0.8.5
numpy>=1.14.5
pandas>=0.23.1
pygbif>=0.2.0
requests>=2.19.1